Replace Elasticsearch facets with aggregations

+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-server/src/main/java/org/graylog2/rest/resources/search/responses/TermsResult.java
@@ -1,18 +1,18 @@
 /**
- * This file is part of Graylog2.
+ * This file is part of Graylog.
  *
- * Graylog2 is free software: you can redistribute it and/or modify
+ * Graylog is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * Graylog2 is distributed in the hope that it will be useful,
+ * Graylog is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with Graylog2.  If not, see <http://www.gnu.org/licenses/>.
+ * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
 package org.graylog2.rest.resources.search.responses;
 
@@ -29,7 +29,7 @@ public abstract class TermsResult {
     public abstract long time();
 
     @JsonProperty
-    public abstract Map<String, Integer> terms();
+    public abstract Map<String, Long> terms();
 
     @JsonProperty
     public abstract long missing();
@@ -44,7 +44,7 @@ public abstract class TermsResult {
     public abstract String builtQuery();
 
     public static TermsResult create(long time,
-                                     Map<String, Integer> terms,
+                                     Map<String, Long> terms,
                                      long missing,
                                      long other,
                                      long total,
