+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-server/src/main/java/org/graylog2/rest/resources/sources/responses/SourcesList.java
@@ -1,18 +1,18 @@
 /**
- * This file is part of Graylog2.
+ * This file is part of Graylog.
  *
- * Graylog2 is free software: you can redistribute it and/or modify
+ * Graylog is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * Graylog2 is distributed in the hope that it will be useful,
+ * Graylog is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with Graylog2.  If not, see <http://www.gnu.org/licenses/>.
+ * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
 package org.graylog2.rest.resources.sources.responses;
 
@@ -29,7 +29,7 @@ public abstract class SourcesList {
     public abstract int total();
 
     @JsonProperty
-    public abstract Map<String, Integer> sources();
+    public abstract Map<String, Long> sources();
 
     @JsonProperty
     public abstract long tookMs();
@@ -37,7 +37,7 @@ public abstract class SourcesList {
     @JsonProperty
     public abstract int range();
 
-    public static SourcesList create(int total, Map<String, Integer> sources, long tookMs, int range) {
+    public static SourcesList create(int total, Map<String, Long> sources, long tookMs, int range) {
         return new AutoValue_SourcesList(total, sources, tookMs, range);
     }
